requirements.txt
PyVISA
pyserial
